requirements.txt
# Core dependencies for RAG experiments

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Embeddings and vector search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
torch>=2.0.0

# LLM interface
openai>=1.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tiktoken>=0.5.0  # OpenAI tokenizer for chunking
tqdm>=4.65.0  # Progress bars

# Evaluation
scikit-learn>=1.3.0

# Data collection (arXiv)
arxiv>=1.4.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
ipython>=8.0.0
jupyter>=1.0.0
